requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.7
Flask-Mail==0.10.0
Werkzeug==3.0.3
WTForms==3.1.2
email-validator==2.2.0
SQLAlchemy==2.0.36
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2
tensorflow==2.17.0
keras==3.5.0
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.3
tqdm==4.66.5
python-socketio==5.11.4
eventlet==0.37.0
python-dotenv==1.0.1
imbalanced-learn==0.12.3
